use a smaller write for large write test on illumos
On illumos systems, the current write of 4096 bytes appears to block
forever waiting for a reader on the far side of the pipe.  The test
mentions that threads will be required to make this work in all cases,
which is generally true, but by dropping the size on illumos to 2048
bytes we can make the test pass without making it more complex.